requirements.txt
GPUtil-fix==1.4.1.2
hid==1.0.7
psutil==7.0.0
pydbus==0.6.0
pyhidapi==0.0.2
